concatenate the LRA only with 12 months

+++ aqua/lra_generator/lra_generator.py
@@ -172,7 +172,7 @@ class LRAgenerator():
         self.reader = Reader(model=self.model, exp=self.exp,
                              source=self.source, zoom=self.zoom,
                              regrid=self.resolution, freq=self.frequency,
-                             configdir=self.configdir, loglevel=self.loglevel)
+                             configdir=self.configdir, loglevel=self.loglevel, rebuild=True)
 
 
         self.logger.info('Accessing catalog for %s-%s-%s...',
@@ -241,6 +241,7 @@ class LRAgenerator():
         cat_file['sources'][entry_name] = block_cat
         dump_yaml(outfile=catalogfile, cfg=cat_file)
 
+
     def _set_dask(self):
         """
         Set up dask cluster
@@ -277,24 +278,25 @@ class LRAgenerator():
     def _concat_var(self, var, year):
         """
         To reduce the amount of files concatenate together all the files
-        from the same year
+        from the same year: do it only if there are 12 files (assuming monthly based data)
         """
 
         infiles =  os.path.join(self.outdir,
                     f'{var}_{self.exp}_{self.resolution}_{self.frequency}_{year}??.nc')
-        xfield = xr.open_mfdataset(infiles)
-        self.logger.warning('Creating a single file for %s, year %s...',  var, str(year))
-        outfile = os.path.join(self.outdir,
-                    f'{var}_{self.exp}_{self.resolution}_{self.frequency}_{year}.nc')
-        # clean older file
-        if os.path.exists(outfile):
-            os.remove(outfile)
-        xfield.to_netcdf(outfile)
+        if len(glob.glob(infiles))==12:
+            xfield = xr.open_mfdataset(infiles)
+            self.logger.warning('Creating a single file for %s, year %s...',  var, str(year))
+            outfile = os.path.join(self.outdir,
+                        f'{var}_{self.exp}_{self.resolution}_{self.frequency}_{year}.nc')
+            # clean older file
+            if os.path.exists(outfile):
+                os.remove(outfile)
+            xfield.to_netcdf(outfile)
 
-        # clean of monthly files
-        for infile in glob.glob(infiles):
-            self.logger.info('Cleaning %s...',  infile)
-            os.remove(infile)
+            # clean of monthly files
+            for infile in glob.glob(infiles):
+                self.logger.info('Cleaning %s...',  infile)
+                os.remove(infile)
 
     def get_filename(self, var, year=None, month=None):
